quart_app/requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
aiohttp-index==0.1
aiohttp-jinja2==1.0.0
appdirs==1.4.3
async-timeout==3.0.0
atomicwrites==1.2.1
attrs==18.2.0
base58==1.0.0
black==18.9b0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
h11==0.8.1
h2==3.0.1
hpack==3.0.0
Hypercorn==0.3.0
hyperframe==5.1.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
json-logging==0.0.9
json-logging-py==0.2
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
multidict==4.4.2
pluggy==0.7.1
py==1.6.0
pytest==3.6.4
pytest-asyncio==0.9.0
python-engineio==2.2.0
python-socketio==2.0.0
python3-indy==1.6.6
pytoml==0.1.19
Quart==0.6.7
quart-openapi==1.3.6
requests==2.20.0
serpy==0.3.1
six==1.11.0
socketIO-client==0.7.2
sortedcontainers==2.0.5
toml==0.9.6
typing-extensions==3.6.5
urllib3==1.23
websocket-client==0.53.0
websockets==6.0
Werkzeug==0.14.1
wsproto==0.12.0
yarl==1.2.6
